requirements.txt
Flask
psycopg2-binary
qrcode
Pillow
pandas
openpyxl
reportlab
python-dotenv